Move e2e tests to a single package
