+++ test/e2e/create_enable_disable_delete_test.go
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package lifecycle
+package e2e_test
 
 import (
 	"context"
